+++ src/bcbench/operations/__init__.py
@@ -16,6 +16,7 @@ from bcbench.operations.git_operations import (
 )
 from bcbench.operations.instruction_operations import copy_problem_statement_folder, setup_custom_agent, setup_instructions_from_config
 from bcbench.operations.project_operations import categorize_projects
+from bcbench.operations.setup_operations import setup_repo
 from bcbench.operations.test_operations import extract_tests_from_patch
 
 __all__ = [
@@ -33,5 +34,6 @@ __all__ = [
     "run_tests",
     "setup_custom_agent",
     "setup_instructions_from_config",
+    "setup_repo",
     "stage_and_get_diff",
 ]
